feat: refactor sidebat to useAlertsDateFilter composable
